Add Python module for testing beat tracking
Hopefully @ervanalb doesn't think it's too ugly to create this new
`beat_tracking_test` binary & `beat_toolkit/` folder.

I had to create a mirror of the Python `rubberband` to fix some minor
bugs in it.

The idea is to resample, mix, and concatenate a few `.wav` files with
known beat timestamps in order to generate a larger corpus of tests.